api/requirements.txt
aiohttp~=3.8.1
alembic~=1.7.4
asyncpg~=0.24.0
cryptography~=35.0.0
databases~=0.5.3
fastapi~=0.70.0
flake8
passlib[bcrypt]~=1.7.4
psycopg2~=2.9.1
pydantic[email]~=1.8.2
pytest~=6.2.5
pytest-asyncio~=0.16.0
pytest-cov~=3.0.0
python-jose~=3.3.0
python-multipart~=0.0.5
requests~=2.26.0
slowapi~=0.1.5
SQLAlchemy~=1.4.25
uvicorn~=0.15.0